+++ orderbook/src/api/handler.rs
@@ -107,7 +107,7 @@ pub async fn get_fee_info(sell_token: H160) -> Result<impl warp::Reply, Infallib
         expiration_date: chrono::offset::Utc::now()
             + FixedOffset::east(STANDARD_VALIDITY_FOR_FEE_IN_SEC),
         minimal_fee: U256::zero(),
-        fee_ratio: 0 as u32,
+        fee_ratio: 0u32,
     };
     Ok(with_status(warp::reply::json(&fee_info), StatusCode::OK))
 }
